added spacing to the profileRows and resolved issues from code review
